Betting-against-beta-strategy/requirements.txt
plotly==4.14.3
streamlit==0.75.0
pandas==1.1.3
numpy==1.19.2
python_dateutil==2.8.1
